+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - `CharacterSet::Pure#keep_in`, `#delete_in` now preserve the original encoding
+- `#count` now supports passing an argument or block
 
 ## [1.2.0] - 2019-04-02